Factored out repertoire	abstract widget	into a more M(VC) design.
The new design uses an abstract model class that can hold widget
state, provides basic ajax functionality, and manages event listeners.

This allows subclasses to implement functionality more like
traditional desktop GUIs, where models are often transient.  Or they
can use an ajax-enabled model, which either serves as a facade to
webservices, a non-visual structure for data displayed in a widget,
or both.

Finally, event binding and triggering are now available on both model
and widget, so it is easier to construct applications out of a
collection of widgets tied together with non-DOM event handlers.